Code I believe is now ready for use.
Copy of convert2ogv.py3 sent to Sameer.
..also rmogv.py incase he wants to do some testing.
scan4html.py3 also seems to be working.

Some constants need to be modified depending on what is to be done.
Specifically the root directory needs set as appropriate.
There are also some options that (1) allow a dummy run
or (2) allow a test run to be done with regard to jpeg files
(since I had some of those around, but not so many with
refs to mp4 files.)
These will be self evident if the constants declared at the
beginning of the code are examined.