Fix pylint violations
To get a 10/10 pylint score, one would need to structurally change the program in a more modular way. However, a reasonable score can be obtained without doing so. For example, convert2ogv.py can have a 8.82/10 score and scan4html.py can have a 9.78/10 score.

This is sufficient as a natural restructuring will over time make these scores rise to 10/10.

There is a line that has no function (a variable is defined but is not used). Commit b967516 makes this code dead and comments this line with phrase FIX. This line is not being used in the rest of the program and will be removed.