Get format_change.py working
Now seems to work except that resulting webm files are reported by
firefox as being corrupted and won't plan.  This has yet to be
investigated and fixed.
format_change.py is a combination of both convert_mp4.py and
scan4html.py: it converts the mp4 files to a specified format AND it
fixes the refs in the html files all in one session.

So yet to do: investigate and fix reason for corrupt webm files.